fix: FP32 norm instead of FP64 — A6000 OOM on .to(float64)
_g.to(torch.float64) doubles the 6.19GB shard to 12.38GB on A6000
(47GB total, only 8GB free after model+shard+grad+Adam states).

Fix: compute norm in FP32 (zero extra allocation since grad is already
FP32), then promote the scalar to float64 for the all_reduce accumulation.
This matches DeepSpeed stage3.py which uses .double() only on the
per-param norm scalars, never on the full gradient buffer.

+++ deepspeed/runtime/desloc_engine.py
@@ -1849,7 +1849,8 @@ class DesLocEngine:
             if self.param_shard_state is not None:
                 _g = self.param_shard_state.param_shard.grad
                 if _g is not None:
-                    local_norm_sq = _g.to(torch.float64).norm(2).pow(2)
+                    # FP32 norm — no extra allocation (grad is already FP32)
+                    local_norm_sq = _g.float().norm(2).to(torch.float64).pow(2)
                 else:
                     local_norm_sq = torch.tensor(0.0, dtype=torch.float64,
                                                   device=self.param_shard_state.param_shard.device)
